requirements.txt
cffi==1.15.0
pony==0.7.14
PyMySQL==1.0.2
python-dotenv==0.19.2
requests==2.26.0
pycparser==2.21
cryptography==36.0.1

